+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Compatibility with WordPress 5.9.
+- New listener for Cavalcade (thanks to [Jan Thiel](https://github.com/JanThiel) for the suggestion).
 
 ### Changed
 - Upgraded Monolog library from version 2.2.0 to version 2.3.4.